Main page: show next event if there are more future ones planned
Previously, the latest planned event was shown on the home page for
each series.
So, when Ostrava Pyvo was planned two months (two events) in advance,
the one in the far future would be shown.

Reuse the logic for ordering series to also apply for picking the event:
prefer a future one if available, then sort by distance from today.

Related to: https://github.com/pyvec/pyvo.cz/issues/86